Visual Studio で Udon.hpp を利用できるように頑張った

+++ src/Udon/Algorithm/ServoSpeed.hpp
@@ -80,7 +80,7 @@ namespace Udon
                                    Udon::Abs(move_angle) * -1, Udon::Abs(move_angle)) *
                         (move_angle < 0 ? -1 : 1);
                 next_angle += value;
-                next_angle = Udon::Constrain(next_angle, 0.0, static_cast<double>(servo_preset));
+                next_angle = Udon::Constrain(next_angle, 0.0f, static_cast<float>(servo_preset));
             }
         }
